Move PasswordField to its own file (for testing).
